fix: コパイロット指摘事項の修正
- userId型変換を統一（Integer → Long）
- CORS設定をワイルドカード(*)から具体的ドメインに変更
- セキュリティ強化とデグレード回避

+++ src/main/java/importApp/mapper/UserMapper.java
@@ -22,6 +22,6 @@ public interface UserMapper {
     int resetFailedLoginAttempts(@Param("username") String username);
     int lockAccount(@Param("username") String username, @Param("lockedUntil") LocalDateTime lockedUntil);
     
-    void updateProfileImageUrl(@Param("userId") Integer userId, @Param("profileImageUrl") String profileImageUrl);
+    void updateProfileImageUrl(@Param("userId") Long userId, @Param("profileImageUrl") String profileImageUrl);
 
 }
